+++ src/main/java/net/meander/subtlyd/data/worldgen/features/FeatureUtilsSD.java
@@ -2,6 +2,7 @@ package net.meander.subtlyd.data.worldgen.features;
 
 import net.meander.subtlyd.util.Util;
 import net.meander.subtlyd.world.level.levelgen.feature.FeatureTypesSD;
+import net.meander.subtlyd.world.level.levelgen.feature.trunkplacers.TrunkPlacerTypeSD;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceKey;
@@ -13,6 +14,7 @@ import net.minecraft.world.level.levelgen.feature.Feature;
 public class FeatureUtilsSD {
     public static void bootstrap() {
         FeatureTypesSD.bootstrap(BuiltInRegistries.FEATURE_TYPE);
+        TrunkPlacerTypeSD.initalize();
     }
 
     public static ResourceKey<Feature> createKey(final String name) {